implemented suggestions from pull request #6
caveat: since I cleaned the data in makeconcepts.tsv the forms in concepts.tsv are clean, while the ones in raw are not. This led to an error, since in #addconcepts we are adding clean keys from concepts.tsv to the concept dictionary, but in #addforms we are querying coarse keys for Parameter_ID. Best workaround I found was to add function cln() which does the same cleaning as the one in makeconcepts.py so the keys are accessed correctly.